+++ geotrek/trekking/serializers.py
@@ -208,7 +208,7 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
                   'web_links', 'is_park_centered', 'disabled_infrastructure',
                   'parking_location', 'relationships', 'points_reference',
                   'gpx', 'kml', 'source', 'type1', 'type2', 'category', 'structure',
-                  'treks', 'parent', 'children', 'previous', 'next') + \
+                  'treks', 'children', 'previous', 'next') + \
             AltimetrySerializerMixin.Meta.fields + \
             ZoningSerializerMixin.Meta.fields + \
             PublishableSerializerMixin.Meta.fields + \